move _settings into connections where it's used, remove gsm vendor
