--- ecoscope_workflows/_settings/__init__.py
@@ -1,1 +1,0 @@
-from ._settings import _Settings as _Settings
