+++ ecoscope_workflows/connections.py
@@ -10,9 +10,34 @@ from typing import (
 )
 
 from pydantic import Field, SecretStr
-from pydantic_settings import SettingsConfigDict
+from pydantic_settings import (
+    BaseSettings,
+    PydanticBaseSettingsSource,
+    PyprojectTomlConfigSettingsSource,
+    SettingsConfigDict,
+)
 
-from ecoscope_workflows._settings import _Settings
+import ecoscope_workflows.config as config
+
+
+class _Settings(BaseSettings):
+    @classmethod
+    def settings_customise_sources(
+        cls,
+        settings_cls: type[BaseSettings],
+        init_settings: PydanticBaseSettingsSource,
+        env_settings: PydanticBaseSettingsSource,
+        dotenv_settings: PydanticBaseSettingsSource,
+        file_secret_settings: PydanticBaseSettingsSource,
+    ) -> tuple[PydanticBaseSettingsSource, ...]:
+        return (
+            init_settings,
+            env_settings,
+            PyprojectTomlConfigSettingsSource(settings_cls, toml_file=config.PATH),
+            # dotenv_settings,
+            # file_secret_settings,
+        )
+
 
 DataConnectionType = TypeVar("DataConnectionType", bound="DataConnection")
 ClientProtocolType = TypeVar("ClientProtocolType")
